refactor: cleanup and review comments
