+++ package.json
@@ -48,8 +48,8 @@
     "compression-webpack-plugin": "^4.0.0",
     "express": "^4.17.1",
     "express-static-gzip": "^2.0.8",
-    "helmet": "^4.1.1",
     "graphql": "^15.4.0",
+    "helmet": "^4.1.1",
     "html-webpack-plugin": "^4.3.0",
     "http-proxy": "^1.18.1",
     "lodash.merge": "^4.6.2",
